requirements.txt
pip~=22.2.2
wheel~=0.37.1
Pillow~=9.2.0
Jinja2~=3.1.2
distlib~=0.3.5
setuptools~=63.2.0
python-dotenv~=0.20.0
certifi~=2022.6.15
packaging~=21.3
pyparsing~=3.0.9
virtualenv~=20.16.3
filelock~=3.8.0
platformdirs~=2.5.2
numpy~=1.23.1
tensorflow-gpu~=2.9.1
uvicorn~=0.18.2
scikit-image~=0.19.3
fastapi~=0.79.0